+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 ### Enhancements
 
-- Rename the `dsym` upload command to `xcode-build` to better reflect the command's purpose.
+- Rename the `dsym` upload command to `xcode-build` to better reflect the command's purpose. [156](https://github.com/bugsnag/bugsnag-cli/pull/156)
+- Add the `xcode-archive` command to upload Xcode archives to Bugsnag. [157](https://github.com/bugsnag/bugsnag-cli/pull/157)
 
 ## 2.7.0 (2024-11-26)